--- virtualTrainer/src/main/java/br/edu/pucsp/virtualTrainer/VirtualTrainerApplication.java
@@ -1,13 +1,0 @@
-package br.edu.pucsp.virtualTrainer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class VirtualTrainerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(VirtualTrainerApplication.class, args);
-	}
-
-}
