--- virtualTrainer/src/main/java/br/edu/pucsp/virtualTrainer/repository/FieldRepository.java
@@ -1,8 +1,0 @@
-package br.edu.pucsp.virtualTrainer.repository;
-
-import br.edu.pucsp.virtualTrainer.model.entity.Field;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FieldRepository extends JpaRepository<Field, Long> {
-
-}
