--- virtualTrainer/src/main/java/br/edu/pucsp/virtualTrainer/repository/TrainerFieldRepository.java
@@ -1,7 +1,0 @@
-package br.edu.pucsp.virtualTrainer.repository;
-
-import br.edu.pucsp.virtualTrainer.model.entity.TrainerField;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TrainerFieldRepository extends JpaRepository<TrainerField, Long> {
-}
