--- virtualTrainer/src/main/java/br/edu/pucsp/virtualTrainer/repository/TrainerRepository.java
@@ -1,12 +1,0 @@
-package br.edu.pucsp.virtualTrainer.repository;
-
-import br.edu.pucsp.virtualTrainer.model.entity.Trainer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-public interface TrainerRepository extends JpaRepository<Trainer, Long> {
-
-    @Query("SELECT t FROM Trainer t WHERE t.name = :name")
-    Trainer findByName(String name);
-
-}
