--- virtualTrainer/src/main/java/br/edu/pucsp/virtualTrainer/service/TrainerService.java
@@ -1,11 +1,0 @@
-package br.edu.pucsp.virtualTrainer.service;
-
-import br.edu.pucsp.virtualTrainer.transport.request.TrainerRequest;
-import br.edu.pucsp.virtualTrainer.transport.response.TrainerResponse;
-
-public interface TrainerService {
-
-    void createTrainer(TrainerRequest request);
-    TrainerResponse findTrainer(String name);
-    void addFields(String certificate);
-}
