+++ virtualTrainer/src/main/java/br/edu/pucsp/virtualtrainer/VirtualTrainerApplication.java
@@ -1,0 +1,13 @@
+package br.edu.pucsp.virtualtrainer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class VirtualTrainerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(VirtualTrainerApplication.class, args);
+	}
+
+}
