+++ virtualTrainer/src/main/java/br/edu/pucsp/virtualtrainer/model/entity/Field.java
@@ -1,0 +1,41 @@
+package br.edu.pucsp.virtualtrainer.model.entity;
+
+import javax.persistence.*;
+
+@Entity
+public class Field {
+
+    @Id()
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column
+    private String name;
+
+    @Column
+    private boolean certificate;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public boolean isCertificate() {
+        return certificate;
+    }
+
+    public void setCertificate(boolean certificate) {
+        this.certificate = certificate;
+    }
+}
