+++ virtualTrainer/src/main/java/br/edu/pucsp/virtualtrainer/repository/FieldRepository.java
@@ -1,0 +1,8 @@
+package br.edu.pucsp.virtualtrainer.repository;
+
+import br.edu.pucsp.virtualtrainer.model.entity.Field;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FieldRepository extends JpaRepository<Field, Long> {
+
+}
