+++ virtualTrainer/src/main/java/br/edu/pucsp/virtualtrainer/repository/TrainerFieldRepository.java
@@ -1,0 +1,7 @@
+package br.edu.pucsp.virtualtrainer.repository;
+
+import br.edu.pucsp.virtualtrainer.model.entity.TrainerField;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TrainerFieldRepository extends JpaRepository<TrainerField, Long> {
+}
