+++ virtualTrainer/src/main/java/br/edu/pucsp/virtualtrainer/repository/TrainerRepository.java
@@ -1,0 +1,12 @@
+package br.edu.pucsp.virtualtrainer.repository;
+
+import br.edu.pucsp.virtualtrainer.model.entity.Trainer;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+
+public interface TrainerRepository extends JpaRepository<Trainer, Long> {
+
+    @Query("SELECT t FROM Trainer t WHERE t.name = :name")
+    Trainer findByName(String name);
+
+}
