+++ virtualTrainer/src/main/java/br/edu/pucsp/virtualtrainer/service/TrainerService.java
@@ -1,0 +1,11 @@
+package br.edu.pucsp.virtualtrainer.service;
+
+import br.edu.pucsp.virtualtrainer.transport.request.TrainerRequest;
+import br.edu.pucsp.virtualtrainer.transport.response.TrainerResponse;
+
+public interface TrainerService {
+
+    void createTrainer(TrainerRequest request);
+    TrainerResponse findTrainer(String name);
+    void addFields(String certificate);
+}
