+++ virtualTrainer/src/main/java/br/edu/pucsp/virtualtrainer/transport/response/TrainerResponse.java
@@ -1,0 +1,16 @@
+package br.edu.pucsp.virtualtrainer.transport.response;
+
+import br.edu.pucsp.virtualtrainer.model.dto.TrainerDto;
+
+public class TrainerResponse {
+
+    private TrainerDto trainer;
+
+    public TrainerDto getTrainer() {
+        return trainer;
+    }
+
+    public void setTrainer(TrainerDto trainer) {
+        this.trainer = trainer;
+    }
+}
