--- virtualTrainer/src/test/java/br/edu/pucsp/virtualTrainer/AbstractRepositoryTest.java
@@ -1,8 +1,0 @@
-package br.edu.pucsp.virtualTrainer;
-
-import org.springframework.test.annotation.Rollback;
-import org.springframework.test.context.junit4.AbstractTransactionalJUnit4SpringContextTests;
-
-@Rollback
-public class AbstractRepositoryTest extends AbstractTransactionalJUnit4SpringContextTests {
-}
