--- virtualTrainer/src/test/java/br/edu/pucsp/virtualTrainer/VirtualTrainerApplicationTests.java
@@ -1,13 +1,0 @@
-package br.edu.pucsp.virtualTrainer;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class VirtualTrainerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
