+++ virtualTrainer/src/test/java/br/edu/pucsp/virtualtrainer/AbstractRepositoryTest.java
@@ -1,0 +1,8 @@
+package br.edu.pucsp.virtualtrainer;
+
+import org.springframework.test.annotation.Rollback;
+import org.springframework.test.context.junit4.AbstractTransactionalJUnit4SpringContextTests;
+
+@Rollback
+public class AbstractRepositoryTest extends AbstractTransactionalJUnit4SpringContextTests {
+}
