+++ virtualTrainer/src/test/java/br/edu/pucsp/virtualtrainer/VirtualTrainerApplicationTests.java
@@ -1,0 +1,13 @@
+package br.edu.pucsp.virtualtrainer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class VirtualTrainerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
